bug fixes for power management controls

+++ orte/include/orte/types.h
@@ -139,7 +139,7 @@ typedef struct {
         opal_byte_object_t bo;
         opal_buffer_t buf;
         float fval;
-        float dval;
+        double dval;
         struct timeval tv;
         void *ptr;  // never packed or passed anywhere
         orte_vpid_t vpid;
